+++ dspot/src/main/java/eu/stamp_project/dspot/assertiongenerator/assertiongenerator_components/testmethodreconstructor_components/AssertionSyntaxBuilder.java
@@ -1,8 +1,8 @@
-package eu.stamp_project.dspot.assertgenerator.components.testmethodreconstructor;
+package eu.stamp_project.dspot.assertiongenerator.assertiongenerator_components.testmethodreconstructor_components;
 
-import eu.stamp_project.dspot.assertgenerator.components.testmethodreconstructor.assertionsyntaxbuilder.AggregateTypeBuilder;
-import eu.stamp_project.dspot.assertgenerator.components.testmethodreconstructor.assertionsyntaxbuilder.InvocationSetup;
-import eu.stamp_project.dspot.assertgenerator.components.testmethodreconstructor.assertionsyntaxbuilder.Translator;
+import eu.stamp_project.dspot.assertiongenerator.assertiongenerator_components.testmethodreconstructor_components.assertionsyntaxbuilder_components.AggregateTypeBuilder;
+import eu.stamp_project.dspot.assertiongenerator.assertiongenerator_components.testmethodreconstructor_components.assertionsyntaxbuilder_components.InvocationSetup;
+import eu.stamp_project.dspot.assertiongenerator.assertiongenerator_components.testmethodreconstructor_components.assertionsyntaxbuilder_components.Translator;
 import eu.stamp_project.test_framework.assertions.AssertEnum;
 import eu.stamp_project.test_framework.TestFramework;
 import eu.stamp_project.utils.TypeUtils;
