+++ dspot/src/test/java/eu/stamp_project/dspot/assertiongenerator/AssertionGeneratorTest.java
@@ -1,8 +1,8 @@
-package eu.stamp_project.dspot.assertgenerator;
+package eu.stamp_project.dspot.assertiongenerator;
 
 import eu.stamp_project.AbstractTest;
 import eu.stamp_project.Utils;
-import eu.stamp_project.dspot.assertgenerator.components.AssertionRemover;
+import eu.stamp_project.dspot.assertiongenerator.assertiongenerator_components.AssertionRemover;
 import eu.stamp_project.test_framework.TestFramework;
 import eu.stamp_project.utils.AmplificationHelper;
 import eu.stamp_project.utils.DSpotUtils;
