+++ dspot/src/test/java/eu/stamp_project/dspot/assertiongenerator/AssertionRemoverTest.java
@@ -1,9 +1,9 @@
-package eu.stamp_project.dspot.assertgenerator;
+package eu.stamp_project.dspot.assertiongenerator;
 
 import eu.stamp_project.AbstractTest;
 import eu.stamp_project.Utils;
-import eu.stamp_project.dspot.assertgenerator.components.AssertionRemover;
-import eu.stamp_project.dspot.assertgenerator.components.testmethodreconstructor.observer.TestWithLogGenerator;
+import eu.stamp_project.dspot.assertiongenerator.assertiongenerator_components.AssertionRemover;
+import eu.stamp_project.dspot.assertiongenerator.assertiongenerator_components.testmethodreconstructor_components.observer_components.TestWithLogGenerator;
 import eu.stamp_project.test_framework.TestFramework;
 import eu.stamp_project.utils.AmplificationHelper;
 import org.junit.BeforeClass;
