+++ dspot/src/test/java/eu/stamp_project/dspot/assertiongenerator/AssertionSyntaxBuilderTest.java
@@ -1,8 +1,8 @@
-package eu.stamp_project.dspot.assertgenerator;
+package eu.stamp_project.dspot.assertiongenerator;
 
 import eu.stamp_project.AbstractTest;
 import eu.stamp_project.Utils;
-import eu.stamp_project.dspot.assertgenerator.components.testmethodreconstructor.AssertionSyntaxBuilder;
+import eu.stamp_project.dspot.assertiongenerator.assertiongenerator_components.testmethodreconstructor_components.AssertionSyntaxBuilder;
 import org.junit.Test;
 import spoon.reflect.code.CtStatement;
 import spoon.reflect.declaration.CtMethod;
@@ -15,7 +15,7 @@ public class AssertionSyntaxBuilderTest extends AbstractTest {
     public void testAssertArrays() {
 
         /*
-            Make sure that observations containing arrays of single or multiple dimensions with components of primitive type
+            Make sure that observations containing arrays of single or multiple dimensions with assertiongenerator_components of primitive type
             and empty arrays correctly generate assertions statements.
          */
